Add command to define word using gptel
Add `my/gptel-define-word` interactive function to define the word or phrase in
the current region using `gptel`.  It uses a new variable,
`my/gptel-word-definition-prompt`, to specify the prompt.  Include
`my/gptel-stash-response` helper function to store the result of the `gptel`
query in a dedicated buffer called "*Last Definition*".  Bind the new command to
`C-c g w` in `global-map`.