Add branch creation and checkout to `my/gptel-generate-branch-name`
Automate branch creation and checkout within `my/gptel-generate-branch-name`.
This adds the ability to immediately create and switch to the new branch.  After
generating and confirming the branch name, prompt the user for the starting
point using `magit-read-branch-or-commit`.  Then, create the branch from the
selected point using `magit-branch-create` and checkout the new branch using
`magit-checkout`.  Update the docstring to reflect the new steps in the
workflow.