Prevent incorrect blank line after `#+TBLFM:`
Correct the `my/org-determine-required-blanks` behavior in
`lisp/org-configuration.el`. This ensures that there is no blank line
after an `#+TBLFM:` metadata line, and also no blank line before it if
the previous content was body text (typically a table).

This change prevents incorrect blank lines, specifically for `#+TBLFM:`,
which previously occurred due to `my/org-determine-required-blanks` not
handling this specific metadata type. This improves the consistency of
`org` file formatting.