Move Lisp configuration to `lisp-configuration.el`
Move the `use-package` declaration for `elisp-autofmt` and the
`emacs-lisp-mode-hook` for `my/add-collapse-to-before-save` from
`init.el` to the new `lisp/lisp-configuration.el` file.

This refactoring centralizes all Lisp and Emacs Lisp-specific
configurations into a dedicated module. It improves the organization and
modularity of the Emacs configuration, making `init.el` cleaner and
easier to manage.